Rounds and Antes (Balabrok.v0.0.0B)
Rounds and Antes are added.

A round is every battle you need to beat that has a determined chips that need to be surpased.
There are theoretically 4 rounds. The 2nd one multiplies by x1.5 the chips of the first, the 3rd one does it by x2 and the 4th one by x3.
An ante is "a difficulty" of the sets of 4 rounds. It gives the chips to beat to the 1st round of the set.
After beating the 4 rounds the Ante will scale up (Ante up), highering the base chips.

Also you can win after you get to the 3rd Ante.